Rename csv script files
