+++ scripts/convert-csv-to-measures.js
@@ -1,0 +1,36 @@
+/**
+ * [convertCsvToMeasures description]
+ * @param  {array of arrays}  records each array in the outer array represents a new measure, each inner array its attributes
+ * @param  {object}           config  object defining how to build a new measure from this csv file, including mapping of measure fields to column indices
+ * @return {array}            Returns an array of measures objects
+ */
+const convertCsvToMeasures = function(records, config) {
+  const sourcedFields = config.sourced_fields;
+  const constantFields = config.constant_fields;
+
+  const newMeasures = records.map(function(record) {
+    var newMeasure = {};
+    Object.entries(sourcedFields).forEach(function([measureKey, columnObject]) {
+      if (typeof columnObject === 'number') {
+        if (!record[columnObject]) {
+          throw TypeError('Column ' + columnObject + ' does not exist in source data');
+        } else {
+          // measure data maps directly to data in csv
+          newMeasure[measureKey] = record[columnObject];
+        }
+      } else {
+        // measure data requires mapping CSV data to new value, e.g. Y, N -> true, false
+        const mappedValue = columnObject.mappings[record[columnObject.index]];
+        newMeasure[measureKey] = mappedValue || columnObject.mappings['default'];
+      }
+    });
+    Object.entries(constantFields).forEach(function([measureKey, measureValue]) {
+      newMeasure[measureKey] = measureValue;
+    });
+    return newMeasure;
+  });
+
+  return newMeasures;
+};
+
+module.exports = convertCsvToMeasures;
